Restructure repo files
