+++ .gitignore
@@ -1,0 +1,5 @@
+.idea/
+__pycache__/
+.ipynb_checkpoints/
+
+env/
